refactor: remove compat shims — delete old flat modules entirely
Remove abicheck/compat.py, abicheck/xml_report.py, abicheck/abicc_dump_import.py.
These were kept as shims but have no external users — all imports are within
the package itself. Tests updated to import from canonical locations:
  abicheck.compat.{descriptor,xml_report,abicc_dump_import}
