--- abicheck/abicc_dump_import.py
@@ -1,2 +1,0 @@
-"""Backward-compatible shim. Implementation moved to abicheck.compat.abicc_dump_import."""
-from .compat.abicc_dump_import import *  # noqa: F403
