--- abicheck/compat.py
@@ -1,2 +1,0 @@
-"""Backward-compatible shim. Implementation moved to abicheck.compat.descriptor."""
-from .compat.descriptor import CompatDescriptor, parse_descriptor  # noqa: F401
