+++ tests/test_xml_parity.py
@@ -21,7 +21,7 @@ import pytest
 from defusedxml.ElementTree import fromstring as xml_fromstring
 
 from abicheck.checker import Change, ChangeKind, DiffResult, Verdict
-from abicheck.xml_report import generate_xml_report
+from abicheck.compat.xml_report import generate_xml_report
 
 # ---------------------------------------------------------------------------
 # Schema validation (unit-level, always runs)
